Implement State for Account, not StorageAccount
