+++ programs/storage_api/src/storage_contract.rs
@@ -59,20 +59,20 @@ pub struct StorageAccount<'a> {
     account: &'a mut Account,
 }
 
-pub trait State<T> {
+trait State<T> {
     fn state(&self) -> Result<T, InstructionError>;
     fn set_state(&mut self, state: &T) -> Result<(), InstructionError>;
 }
 
-impl<'a, T> State<T> for StorageAccount<'a>
+impl<T> State<T> for Account
 where
     T: serde::Serialize + serde::de::DeserializeOwned,
 {
     fn state(&self) -> Result<T, InstructionError> {
-        deserialize(&self.account.data).map_err(|_| InstructionError::InvalidAccountData)
+        deserialize(&self.data).map_err(|_| InstructionError::InvalidAccountData)
     }
     fn set_state(&mut self, state: &T) -> Result<(), InstructionError> {
-        serialize_into(&mut self.account.data[..], state).map_err(|err| match *err {
+        serialize_into(&mut self.data[..], state).map_err(|err| match *err {
             ErrorKind::SizeLimit => InstructionError::AccountDataTooSmall,
             _ => InstructionError::GenericError,
         })
@@ -91,7 +91,7 @@ impl<'a> StorageAccount<'a> {
         entry_height: u64,
         signature: Signature,
     ) -> Result<(), InstructionError> {
-        let mut storage_contract = &mut self.state()?;
+        let mut storage_contract = &mut self.account.state()?;
         if let StorageContract::Default = storage_contract {
             *storage_contract = StorageContract::ReplicatorStorage {
                 proofs: vec![],
@@ -121,7 +121,7 @@ impl<'a> StorageAccount<'a> {
                 signature,
             };
             proofs[segment_index] = proof_info;
-            self.set_state(storage_contract)
+            self.account.set_state(storage_contract)
         } else {
             Err(InstructionError::InvalidArgument)?
         }
@@ -132,7 +132,7 @@ impl<'a> StorageAccount<'a> {
         hash: Hash,
         entry_height: u64,
     ) -> Result<(), InstructionError> {
-        let mut storage_contract = &mut self.state()?;
+        let mut storage_contract = &mut self.account.state()?;
         if let StorageContract::Default = storage_contract {
             *storage_contract = StorageContract::ValidatorStorage {
                 entry_height: 0,
@@ -166,7 +166,7 @@ impl<'a> StorageAccount<'a> {
             *reward_validations = lockout_validations.clone();
             lockout_validations.clear();
             lockout_validations.resize(segments as usize, Vec::new());
-            self.set_state(storage_contract)
+            self.account.set_state(storage_contract)
         } else {
             Err(InstructionError::InvalidArgument)?
         }
@@ -178,7 +178,7 @@ impl<'a> StorageAccount<'a> {
         proofs: Vec<CheckedProof>,
         replicator_accounts: &mut [StorageAccount],
     ) -> Result<(), InstructionError> {
-        let mut storage_contract = &mut self.state()?;
+        let mut storage_contract = &mut self.account.state()?;
         if let StorageContract::Default = storage_contract {
             *storage_contract = StorageContract::ValidatorStorage {
                 entry_height: 0,
@@ -202,12 +202,16 @@ impl<'a> StorageAccount<'a> {
             let mut previous_proofs = replicator_accounts
                 .iter_mut()
                 .filter_map(|account| {
-                    account.state().ok().map(move |contract| match contract {
-                        StorageContract::ReplicatorStorage { proofs, .. } => {
-                            Some((account, proofs[segment_index].clone()))
-                        }
-                        _ => None,
-                    })
+                    account
+                        .account
+                        .state()
+                        .ok()
+                        .map(move |contract| match contract {
+                            StorageContract::ReplicatorStorage { proofs, .. } => {
+                                Some((account, proofs[segment_index].clone()))
+                            }
+                            _ => None,
+                        })
                 })
                 .flatten()
                 .collect::<Vec<_>>();
@@ -246,7 +250,7 @@ impl<'a> StorageAccount<'a> {
 
             // allow validators to store successful validations
             lockout_validations[segment_index].append(&mut valid_proofs);
-            self.set_state(storage_contract)
+            self.account.set_state(storage_contract)
         } else {
             Err(InstructionError::InvalidArgument)?
         }
@@ -257,7 +261,7 @@ impl<'a> StorageAccount<'a> {
         entry_height: u64,
         tick_height: u64,
     ) -> Result<(), InstructionError> {
-        let mut storage_contract = &mut self.state()?;
+        let mut storage_contract = &mut self.account.state()?;
         if let StorageContract::Default = storage_contract {
             Err(InstructionError::InvalidArgument)?
         };
@@ -271,7 +275,7 @@ impl<'a> StorageAccount<'a> {
             // TODO can't just create lamports out of thin air
             // self.account.lamports += TOTAL_VALIDATOR_REWARDS * num_validations;
             reward_validations.clear();
-            self.set_state(storage_contract)
+            self.account.set_state(storage_contract)
         } else if let StorageContract::ReplicatorStorage {
             reward_validations, ..
         } = &mut storage_contract
@@ -295,7 +299,7 @@ impl<'a> StorageAccount<'a> {
             //     * TOTAL_REPLICATOR_REWARDS
             //     * (num_validations / reward_validations[claims_index].len() as u64);
             reward_validations.clear();
-            self.set_state(storage_contract)
+            self.account.set_state(storage_contract)
         } else {
             Err(InstructionError::InvalidArgument)?
         }
@@ -308,7 +312,7 @@ fn store_validation_result(
     segment_index: usize,
     status: ProofStatus,
 ) -> Result<(), InstructionError> {
-    let mut storage_contract = storage_account.state()?;
+    let mut storage_contract = storage_account.account.state()?;
     match &mut storage_contract {
         StorageContract::ReplicatorStorage {
             proofs,
@@ -326,7 +330,7 @@ fn store_validation_result(
         }
         _ => return Err(InstructionError::InvalidAccountData),
     }
-    storage_account.set_state(&storage_contract)
+    storage_account.account.set_state(&storage_contract)
 }
 
 fn count_valid_proofs(proofs: &[CheckedProof]) -> u64 {
@@ -337,4 +341,44 @@ fn count_valid_proofs(proofs: &[CheckedProof]) -> u64 {
         }
     }
     num
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_account_data() {
+        solana_logger::setup();
+        let mut account = Account::default();
+        account.data.resize(4 * 1024, 0);
+        let storage_account = StorageAccount::new(&mut account);
+        // pretend it's a validator op code
+        let mut contract = storage_account.account.state().unwrap();
+        if let StorageContract::ValidatorStorage { .. } = contract {
+            assert!(true)
+        }
+        if let StorageContract::ReplicatorStorage { .. } = &mut contract {
+            panic!("this shouldn't work");
+        }
+
+        contract = StorageContract::ValidatorStorage {
+            entry_height: 0,
+            hash: Hash::default(),
+            lockout_validations: vec![],
+            reward_validations: vec![],
+        };
+        storage_account.account.set_state(&contract).unwrap();
+        if let StorageContract::ReplicatorStorage { .. } = contract {
+            panic!("this shouldn't work");
+        }
+        contract = StorageContract::ReplicatorStorage {
+            proofs: vec![],
+            reward_validations: vec![],
+        };
+        storage_account.account.set_state(&contract).unwrap();
+        if let StorageContract::ValidatorStorage { .. } = contract {
+            panic!("this shouldn't work");
+        }
+    }
 }
